chore: resolve code duplication in API views
